% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biginteger.R
\name{formatN.default}
\alias{formatN.default}
\title{Default Method: Use the standard format() --- e.g. for complex}
\usage{
\method{formatN}{default}(x, ...)
}
\description{
Default Method: Use the standard format() --- e.g. for complex
}
