% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigInteger.R
\name{formatN}
\alias{formatN}
\title{format() Numbers such as to distinguish  bigz, integer, double, mpfr, etc}
\usage{
formatN(x, ...)
}
\description{
format() Numbers such as to distinguish  bigz, integer, double, mpfr, etc
}
